@@ -15,25 +15,26 @@
     "typecheck": "tsc"
   },
   "dependencies": {
-    "@docusaurus/core": "3.0.0-beta.0",
-    "@docusaurus/plugin-ideal-image": "^3.0.0-beta.0",
-    "@docusaurus/plugin-pwa": "^3.0.0-beta.0",
-    "@docusaurus/preset-classic": "3.0.0-beta.0",
+    "@docusaurus/core": "3.0.0",
+    "@docusaurus/plugin-ideal-image": "^3.0.0",
+    "@docusaurus/plugin-pwa": "^3.0.0",
+    "@docusaurus/preset-classic": "^3.0.0",
     "@giscus/react": "^2.3.0",
-    "@mdx-js/react": "^2.3.0",
+    "@mdx-js/react": "^3.0.0",
     "clsx": "^1.2.1",
     "docusaurus-plugin-sass": "^0.2.5",
     "path-browserify": "^1.0.1",
     "plugin-image-zoom": "github:flexanalytics/plugin-image-zoom",
-    "prism-react-renderer": "^1.3.5",
+    "prism-react-renderer": "^2.1.0",
     "raw-loader": "^4.0.2",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
-    "sass": "^1.69.3"
+    "sass": "^1.69.5"
   },
   "devDependencies": {
-    "@docusaurus/module-type-aliases": "3.0.0-beta.0",
-    "@docusaurus/tsconfig": "3.0.0-beta.0",
+    "@docusaurus/module-type-aliases": "^3.0.0",
+    "@docusaurus/tsconfig": "^3.0.0",
+    "@docusaurus/types": "^3.0.0",
     "typescript": "~5.2.2"
   },
   "browserslist": {
@@ -49,6 +50,6 @@
     ]
   },
   "engines": {
-    "node": ">=16.14"
+    "node": ">=18.0"
   }
 }
